Add functionality for submitting message
Submitting a message now appends the 'messages' array with a new
object; the user and text properties of which are populated by the
value of the Name and Message inputs, respectively, and creates a new
Date object for the added property value. Also redirects the user to
the '/' route so their new message can be viewed.